extract duplicated code
