+++ src/Framework/WithTestSuite.php
@@ -9,77 +9,12 @@
  */
 namespace PHPUnit\Framework;
 
-use Generator;
-use ReflectionClass;
-
-class WithTestSuite extends TestSuite
+class WithTestSuite extends DataProvidedTestSuite
 {
-    private $theClass;
-    private $method;
-    public function __construct(ReflectionClass $theClass, string $method)
+    protected function yieldData(): iterable
     {
-        parent::__construct($theClass, $theClass->getName().'::'.$method, true);
-        $this->theClass = $theClass;
-        $this->method = $method;
-    }
-    /**
-     * @var string[]
-     */
-    private $dependencies = [];
-
-    /**
-     * @param string[] $dependencies
-     */
-    public function setDependencies(array $dependencies): void
-    {
-        $this->dependencies = $dependencies;
-
-        foreach ($this->tests as $test) {
-            $test->setDependencies($dependencies);
-        }
-    }
-
-    public function getDependencies(): array
-    {
-        return $this->dependencies;
-    }
-
-    public function hasDependencies(): bool
-    {
-        return \count($this->dependencies) > 0;
-    }
-
-    /**
-     * @return Generator|Test[]
-     */
-    protected function yieldTests(): Generator
-    {
-        yield from parent::yieldTests();
-        try {
-            foreach (\PHPUnit\Util\Test::getDataFromTestWithAnnotation($this->theClass->getMethod($this->method)->getDocComment())[0]??[] as $name => $set) {
-                if(!is_array($set)) {
-                    yield self::warning("set $name is invalid.");
-                    continue;
-                }
-                try {
-                    $test = $this->theClass->newInstanceArgs([
-                        $this->method,
-                        $set,
-                        $name
-                    ]);
-                    $test->setDependencies($this->dependencies);
-                    yield $test;
-                } catch (\Throwable $e) {
-                    yield self::warning("Test creation failed. $e");
-                }
-            }
-        } catch (\Throwable $e) {
-            yield self::warning("data provider failed. $e");
-        }
-    }
-
-    public function count($preferCache = false): int
-    {
-        return 1;
+        return \PHPUnit\Util\Test::getDataFromTestWithAnnotation(
+            $this->theClass->getMethod($this->method)->getDocComment()
+        )[0]??[];
     }
 }
